src/xiyan_mcp_server/pyproject.toml
[project]
name = "xiyan-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index>=0.12.47",
    "loguru>=0.7.3",
    "mcp>=1.10.1",
    "mysql>=0.0.3",
    "mysql-connector-python>=9.3.0",
    "openai>=1.93.1",
    "pandas>=2.3.1",
    "pymysql>=1.1.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
]
